+++ tests/files_test.py
@@ -1,43 +1,28 @@
 from tradingbot.paths import config_path
-import tradingbot.files as files
+import tradingbot.files as f
+from tradingbot.files import Files
 from pathlib import Path
 import tempfile
 import os
-from unittest import mock
-
-
-def test_get_successful_symbols(tmp_path):
-
-  with open(tmp_path / 'successful_symbols.txt', 'w') as file:
-    file.write('EURUSD\nUSDJPY\nGBPUSD')
-
-  # Mock data_path to return the temporary file path
-  with mock.patch('tradingbot.files.data_path') as mock_data_path:
-    mock_data_path.return_value = tmp_path
-
-    # Call the function to test
-    result = files.get_successful_symbols()
-
-  # Verify the result
-  assert result == ['EURUSD', 'USDJPY', 'GBPUSD']
+from unittest.mock import patch
 
 
 def test_file_exists():
   with tempfile.NamedTemporaryFile() as temp_file:
     # File exists
     temp_directory = Path(tempfile.mkstemp()[1])
-    assert files.file_exists(temp_file.name, temp_directory)
+    assert f.file_exists(temp_file.name, temp_directory)
 
   # File does not exist
-  assert not files.file_exists('invalid.txt', config_path())
+  assert not f.file_exists('invalid.txt', config_path())
 
 
 def test_try_load_json():
   with tempfile.NamedTemporaryFile() as temp_file:
     file = Path(temp_file.name)
-    with open(file, 'w') as f:
-      f.write('{"test": "test"}')
-    data = files.try_load_json(file)
+    with open(file, 'w') as fi:
+      fi.write('{"test": "test"}')
+    data = f.try_load_json(file)
     assert data['test'] == 'test'
 
 
@@ -45,15 +30,17 @@ def test_try_read_file():
   with tempfile.TemporaryDirectory() as tmp_dir:
 
     # Create the temporary file inside the temporary directory
-    temp_file = Path(tmp_dir) / "temp.txt"
-    with open(temp_file, 'w') as f:
-      f.write('test')
+    filename = "temp.txt"
+    tmp_dir = Path(tmp_dir)
+    temp_file = Path(tmp_dir) / filename
+    with open(temp_file, 'w') as fi:
+      fi.write('test')
 
-    assert 'test' == files.try_read_file(temp_file)
+    assert 'test' == f.try_read_file(filename, tmp_dir)
 
     os.remove(temp_file)
 
-    assert '' == files.try_read_file(temp_file)
+    assert '' == f.try_read_file(filename, tmp_dir)
 
 
 def test_try_remove_file():
@@ -67,7 +54,113 @@ def test_try_remove_file():
     assert temp_file.exists()
 
     # Try to remove it
-    assert files.try_remove_file(temp_file)
+    assert f.try_remove_file(temp_file)
 
     # Validate that it no longer exists
     assert not temp_file.exists()
+
+
+def test_write_file(tmp_path):
+  file_path = Path(tmp_path)
+  f.write_file('test.txt', 'test', file_path=file_path)
+  with open(file_path / 'test.txt', 'r') as fi:
+    assert fi.read() == 'test'
+
+
+@patch('tradingbot.files.data_path')
+def test_lock(mock_data_path, tmp_path):
+
+  # Make data_path() return the temporary directory
+  mock_data_path.return_value = tmp_path
+
+  f.lock(Files.FOREX_LOCK)
+
+  # Check that the file was written
+  assert os.path.exists(tmp_path / Files.FOREX_LOCK.value)
+
+  f.unlock(Files.FOREX_LOCK)
+
+  # Check that the file was removed
+  assert not os.path.exists(tmp_path / Files.FOREX_LOCK.value)
+
+
+@patch('tradingbot.files.data_path')
+def test_reset_successful_symbols_file(mock_data_path, tmp_path):
+
+  # Make data_path() return the temporary directory
+  mock_data_path.return_value = tmp_path
+
+  # Create the file with some content
+  mock_text = 'EURUSD\nUSDJPY\nGBPUSD'
+  with open(tmp_path / Files.SUCCESSFUL_SYMBOLS, 'w') as file:
+    file.write(mock_text)
+
+  assert f.try_read_file(Files.SUCCESSFUL_SYMBOLS) == mock_text
+
+  # Call to the function to test
+  f.reset_successful_symbols_file()
+
+  assert f.try_read_file(Files.SUCCESSFUL_SYMBOLS) == ''
+
+
+@patch('tradingbot.files.data_path')
+def test_reset_consecutive_times_down_file(mock_data_path, tmp_path):
+
+  # Make data_path() return the temporary directory
+  mock_data_path.return_value = tmp_path
+
+  # Create the file with some content
+  mock_text = '10'
+  with open(tmp_path / Files.CONSECUTIVE_TIMES_DOWN, 'w') as file:
+    file.write(mock_text)
+
+  assert f.try_read_file(Files.CONSECUTIVE_TIMES_DOWN) == mock_text
+
+  # Call to the function to test
+  f.reset_consecutive_times_down_file()
+
+  assert f.try_read_file(Files.CONSECUTIVE_TIMES_DOWN) == '0'
+
+
+@patch('tradingbot.files.data_path')
+def test_get_last_balance(mock_data_path, tmp_path):
+
+  # Make data_path() return the temporary directory
+  mock_data_path.return_value = tmp_path
+
+  # Create the file with some content
+  with open(tmp_path / Files.LAST_BALANCE, 'w') as file:
+    file.write('100')
+
+  # Call to the function to test
+  assert f.get_last_balance() == 100
+
+
+@patch('tradingbot.files.data_path')
+def test_get_consecutive_times_down(mock_data_path, tmp_path):
+
+  # Make data_path() return the temporary directory
+  mock_data_path.return_value = tmp_path
+
+  # Create the file with some content
+  with open(tmp_path / Files.CONSECUTIVE_TIMES_DOWN, 'w') as file:
+    file.write('10')
+
+  # Call to the function to test
+  assert f.get_consecutive_times_down() == 10
+
+
+@patch('tradingbot.files.data_path')
+def test_increment_consecutive_times_down(mock_data_path, tmp_path):
+
+  # Make data_path() return the temporary directory
+  mock_data_path.return_value = tmp_path
+
+  # Create the file with some content
+  with open(tmp_path / Files.CONSECUTIVE_TIMES_DOWN, 'w') as file:
+    file.write('10')
+
+  # Call to the function to test
+  f.increment_consecutive_times_down()
+
+  assert f.get_consecutive_times_down() == 11