+++ tests/forex_test.py
@@ -1,15 +1,22 @@
 from tradingbot.forex import is_locked, check_time_viability
-from tradingbot.paths import config_path
+from tradingbot.files import Files
 from freezegun import freeze_time
 from datetime import datetime
+from pathlib import Path
 from tradingbot.config import Config
+from unittest.mock import patch
 import pytz
 
 
-def test_is_locked():
-  # Test when lock file exists
-  lock_file = config_path() / 'forex_lock'
+@patch('tradingbot.files.data_path')
+def test_is_locked(mock_data_path, tmp_path):
+
+  # Make data_path() return the temporary directory
+  mock_data_path.return_value = tmp_path
+
+  lock_file = Path(tmp_path / Files.FOREX_LOCK.value)
   lock_file.touch()
+
   assert is_locked()
 
   # Test when lock file does not exist
@@ -18,7 +25,7 @@ def test_is_locked():
 
 
 def test_check_time_viability():
-  tz = pytz.timezone(str(Config.forex_timezone))
+  tz = pytz.timezone(str(Config.broker_timezone))
 
   # Correct
   d = datetime(2024, 1, 9, 12, 5)
@@ -26,7 +33,7 @@ def test_check_time_viability():
     assert check_time_viability()
 
   # Weekday
-  d = datetime(2024, 1, 14, 12, 5)
+  d = datetime(2024, 1, 27, 12, 5)
   with freeze_time(tz.localize(d)):
     assert not check_time_viability()
 
@@ -35,12 +42,12 @@ def test_check_time_viability():
   with freeze_time(tz.localize(d)):
     assert not check_time_viability()
 
-  # Friday after 17:00
-  d = datetime(2024, 1, 12, 17, 5)
+  # Friday after 00:00
+  d = datetime(2024, 1, 27, 00, 5)
   with freeze_time(tz.localize(d)):
     assert not check_time_viability()
 
-  # Sunday before 17:00
-  d = datetime(2024, 1, 14, 16, 5)
+  # Sunday before 00:00
+  d = datetime(2024, 1, 28, 23, 55)
   with freeze_time(tz.localize(d)):
     assert not check_time_viability()