+++ tests/log_test.py
@@ -1,4 +1,4 @@
-from tradingbot.log import Log, log
+from tradingbot.log import log
 
 
 def test_info():
@@ -19,16 +19,3 @@ def test_warning():
 def test_error():
   log.error('Testing error')
   assert True
-
-
-def test_singleton():
-  log1 = Log()
-  log2 = Log()
-
-  class NotSingleton():
-    pass
-  ns1 = NotSingleton()
-  ns2 = NotSingleton()
-
-  assert id(log1) == id(log2)
-  assert id(ns1) != id(ns2)