+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 # CHANGELOG
 
+## ** V0.3.0  (25/01/2024) **
+- feat:
+  - forex main to handle incoming historic data
+  - support methods
+- test
+  - forex main
+
 ## ** V0.2.0  (18/01/2024) **
 - feat:
   - agent in MQL5
   - files utils
   - new util method
+  - new config parameters
 - test
   - files utils
+  - agent in MQL
+  - utils
   - forex client
-  - new config parameters
 
 ## ** V0.1.0  (05/01/2024) **
 - feat: